INTG-2353 refactor
